relocated load() close() get_driver() and screenshot() to basePage.py
