+++ pages/visaPage.py
@@ -17,10 +17,7 @@ class VisaPage(BasePage):
     SUBMIT_BTN = (By.ID, "submit")
     TODAY = datetime.today().strftime('%d-%m-%Y')
 
-    def load(self):
-        self.browser.get(VisaPage.URL)
-        self.browser.maximize_window()
-    
+   
     def countriesFromTo(self,VISA_FROM_COUNTRY, VISA_TO_COUNTRY):
         self.browser.find_element(*VisaPage.FROM_COUNTRY_DROPDOWN).click()
         self.browser.find_element(*VisaPage.COUNTRY_INPUT).send_keys(VISA_FROM_COUNTRY)
@@ -29,13 +26,6 @@ class VisaPage(BasePage):
         self.browser.find_element(*VisaPage.COUNTRY_INPUT).send_keys(VISA_TO_COUNTRY)
         self.browser.find_element(*VisaPage.COUNTRY_INPUT).send_keys(Keys.ENTER)
         self.browser.find_element(*VisaPage.SUBMIT_BTN).click()
-
-    def close(self):
-        self.browser.close()
-    
-    def screenshot(self):
-        date_time = time.strftime("%a.%d.%b.%Y.%H%M%S")
-        self.browser.save_screenshot(f"resources\\screenshots\\{date_time}.png")
 
     def getAttrValue(self, element, atribut):
         return self.browser.find_element(*element).get_attribute(atribut)
