+++ tests/test_home.py
@@ -11,8 +11,10 @@ class Test_home_page():
         self.home_page = HomePage(ConfigC)
 
     def test_link_checks(self):
-        self.home_page.load()
-        self.home_page.link_checker()
+        self.home_page.load(self.home_page.URL)
+        results = self.home_page.link_checker()
+        assert results == True, "One or more bad link(s) found"
+
  
     def teardown(self):
         self.home_page.close()
